feat: add /api/v1/profile GET and PATCH endpoints
