+++ app/api/v1/profile/route.ts
@@ -1,0 +1,27 @@
+import { getCredentialAccountHandler } from '@/lib/services/accounts/credentialsHandler'
+import { updateCredentialsHandler } from '@/lib/services/accounts/updateCredentialsHandler'
+import { HttpMethod } from '@/lib/utils/http-headers'
+import { defaultOptions } from '@/lib/utils/response'
+import { traceApiRoute } from '@/lib/utils/traceApiRoute'
+
+const CORS_HEADERS = [
+  HttpMethod.enum.OPTIONS,
+  HttpMethod.enum.GET,
+  HttpMethod.enum.PATCH
+]
+
+export const OPTIONS = defaultOptions(CORS_HEADERS)
+
+// GET /api/v1/profile — view the current actor's CredentialAccount.
+// https://docs.joinmastodon.org/methods/profile/#get
+// Scope: read:accounts (satisfied by the aggregate `read`).
+export const GET = traceApiRoute(
+  'getProfile',
+  getCredentialAccountHandler(CORS_HEADERS)
+)
+
+// PATCH /api/v1/profile — update the current actor's profile and return the
+// updated CredentialAccount. Same semantics as update_credentials.
+// https://docs.joinmastodon.org/methods/profile/#update
+// Scope: write:accounts (satisfied by the aggregate `write`).
+export const PATCH = traceApiRoute('patchProfile', updateCredentialsHandler)
